pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clarence"
version = "0.3.0"
description = "AI day trading execution agent."
authors = [{name = "phin3has"}]
requires-python = ">=3.10"
dependencies = [
    "alpaca-py>=0.32.1",
    "anthropic>=0.40.0",
    "mcp[cli]>=1.0.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.scripts]
clarence = "clarence.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
